chat-client/requirements.txt
python-socketio[client]
openai
google-cloud-aiplatform